fixed admin login,signup,verify and refresh token routes

+++ security/auth.py
@@ -1,11 +1,14 @@
 # auth.py
 from fastapi import Depends, HTTPException, status
 from fastapi.security import HTTPBearer
-token_auth_scheme = HTTPBearer()
-from security.tokens import validate_admin_accesstoken,generate_refresh_tokens,generate_member_access_tokens, validate_member_accesstoken, validate_refreshToken,validate_member_accesstoken_without_expiration
-from repositories.tokens_repo import delete_access_token
-from schemas.tokens_schema import refreshedToken
 
+from security.tokens import validate_admin_accesstoken,generate_refresh_tokens,generate_member_access_tokens, validate_member_accesstoken, validate_refreshToken,validate_member_accesstoken_without_expiration,generate_admin_access_tokens
+from security.encrypting_jwt import decode_jwt_token
+from repositories.tokens_repo import delete_access_token,update_admin_access_tokens
+from schemas.tokens_schema import refreshedToken,accessTokenOut
+
+
+token_auth_scheme = HTTPBearer()
 async def verify_token(token: str = Depends(token_auth_scheme)):
     result = await validate_member_accesstoken(accessToken=token.credentials)
     
@@ -20,28 +23,54 @@ async def verify_token(token: str = Depends(token_auth_scheme)):
         
         
 async def verify_token_and_refresh_token(token: str = Depends(token_auth_scheme)):
-    result = await validate_member_accesstoken_without_expiration(accessToken=token.credentials)
-    accessTokenObj= await generate_member_access_tokens(userId=result.userId)
-    
-    refreshTokenObj = await generate_refresh_tokens(userId=result.userId,accessToken=accessTokenObj.accesstoken)
-    await delete_access_token(result.accesstoken)
-    refreshedTokens = refreshedToken(userId=result.userId,refreshToken=refreshTokenObj.refreshtoken,accessToken=accessTokenObj.accesstoken)
+    decodedT= await decode_jwt_token(token=token.credentials)
+    if decodedT['role']=="member":
+        result = await validate_member_accesstoken_without_expiration(accessToken=token.credentials)
+        
+        accessTokenObj= await generate_member_access_tokens(userId=result.userId)
+        
+        refreshTokenObj = await generate_refresh_tokens(userId=result.userId,accessToken=accessTokenObj.accesstoken)
+        await delete_access_token(result.accesstoken)
+        refreshedTokens = refreshedToken(userId=result.userId,refreshToken=refreshTokenObj.refreshtoken,accessToken=accessTokenObj.accesstoken)
+        if result==None:
+            raise HTTPException(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                detail="Invalid token"
+            )
+        else:
+            return refreshedTokens
+    elif decodedT['role']=="admin":
+        result = await validate_admin_accesstoken(accessToken=str(token.credentials))
+        if result =="inactive":
+            raise HTTPException(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                detail="You Can't Make Use of an Inactive AccessToken"
+            )
+        elif type(result) ==type(accessTokenOut(userId="sa",accesstoken="sa",)):
+            accessTokenObj = await generate_admin_access_tokens(userId=result.userId)
+            # TODO: decode then update to active here 
+            NewDecodedT = await decode_jwt_token(token=accessTokenObj.accesstoken)
+            
+            await update_admin_access_tokens(token=NewDecodedT['accessToken'])
+            refreshTokenObj = await generate_refresh_tokens(userId=result.userId,accessToken=accessTokenObj.accesstoken)
+            await delete_access_token(result.accesstoken)
+            refreshedTokens = refreshedToken(userId=result.userId,refreshToken=refreshTokenObj.refreshtoken,accessToken=accessTokenObj.accesstoken)
+            return refreshedTokens
+        elif result==None:
+            raise HTTPException(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                detail="Invalid Token"
+            )
+           
+async def verify_admin_token(token: str = Depends(token_auth_scheme)):
+    result = await validate_admin_accesstoken(accessToken=str(token.credentials))
     if result==None:
         raise HTTPException(
             status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Invalid token"
+            detail="Invalid admin token"
         )
-    else:
-        return refreshedTokens
-    
-        
-
-
-
-def verify_admin_token(token: str = Depends(token_auth_scheme)):
-    print(token)
-    if token.credentials != "secret-token-admin":
+    elif result=="inactive":
         raise HTTPException(
             status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Invalid admin token"
+            detail="Admin Token hasn't been activated"
         )
